Move patch_urls method in odc.stac._load (23)
This completes decoupling of odc.stac from odc.index.
