+++ odc/index/__init__.py
@@ -14,7 +14,6 @@ from ._index import (
     month_range,
     ordered_dss,
     parse_doc_stream,
-    patch_urls,
     product_from_yaml,
     season_range,
     time_range,
@@ -39,7 +38,6 @@ __all__ = (
     "chopped_dss",
     "all_datasets",
     "product_from_yaml",
-    "patch_urls",
     "odc_uuid",
     "utm_region_code",
     "utm_zone_to_epsg",
