+++ odc/index/_index.py
@@ -4,7 +4,7 @@ import json
 import sys
 from random import randint
 from types import SimpleNamespace
-from typing import Callable, Iterable, Iterator, Optional, Set, Tuple
+from typing import Iterator, Optional, Set, Tuple
 from uuid import UUID
 from warnings import warn
 
@@ -14,7 +14,6 @@ from datacube.api.query import Query
 from datacube.index.hl import Doc2Dataset
 from datacube.index.index import default_metadata_type_docs
 from datacube.model import Dataset, DatasetType, Range, metadata_from_doc
-from datacube.storage import measurement_paths
 from datacube.utils.documents import load_documents, parse_yaml
 from pandas import Period
 
@@ -422,27 +421,3 @@ def product_from_yaml(path: str, dc: Optional[Datacube] = None) -> DatasetType:
         dc = Datacube()
 
     return dc.index.products.from_doc(product)
-
-
-def patch_urls(
-    ds: Dataset, edit: Callable[[str], str], bands: Optional[Iterable[str]] = None
-) -> Dataset:
-    """
-    Map function over dataset measurement urls.
-
-    :param ds: Dataset to edit in place
-    :param edit: Function that returns modified url from input url
-    :param bands: Only edit specified bands, default is to edit all
-    :return: Input dataset
-    """
-    resolved_paths = measurement_paths(ds)
-    if bands is None:
-        bands = list(resolved_paths)
-    else:
-        # remap aliases if present to their canonical name
-        bands = list(map(ds.type.canonical_measurement, bands))
-
-    mm = ds.metadata_doc["measurements"]
-    for band in bands:
-        mm[band]["path"] = edit(resolved_paths[band])
-    return ds
